requirements.txt
accelerate==1.11.0
appdirs==1.4.4
atomicwrites==1.4.0
av==15.1.0
charset-normalizer==2.1.1
cmake==4.1.2
datasets==4.0.0
deepdiff==8.6.1
diffusers==0.35.2
draccus==0.10.0
einops==0.8.1
Farama-Notifications==0.0.4
feetech-servo-sdk==1.0.0
gymnasium==1.2.1
hf-xet==1.1.8
hf_transfer==0.1.9
huggingface-hub==0.34.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
inquirerpy==0.3.4
jiter==0.10.0
jsonlines==4.0.0
jsonpointer==2.1
matplotlib==3.9.2
mergedeep==1.3.4
multiprocess==0.70.16
openai==1.99.9
opencv-python-headless==4.12.0.88
orderly-set==5.5.0
packaging==25.0
pdfminer.six==20250506
pfzy==0.3.4
protobuf==4.25.3
python-dotenv==1.0.1
pyarrow==22.0.0
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
pynput==1.8.1
pypdf==6.1.1
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyserial==3.5
pyyaml-include==1.4.1
rerun-sdk==0.26.2
sacremoses==0.1.1
safetensors==0.6.2
sentry-sdk==2.42.1
setuptools==75.1.0
sympy==1.14.0
termcolor==3.2.0
tqdm==4.66.5
tokenizers==0.21.4
transformers==4.55.2
typing-inspect==0.9.0
unzip==1.0.0
wandb==0.21.4
wheel==0.44.0
xxhash==3.5.0
